+++ src/main/kotlin/com/nylas/models/EventResource.kt
@@ -9,4 +9,14 @@ data class EventResource(
   val email: String,
   @Json(name = "name")
   val name: String? = null,
+  @Json(name = "capacity")
+  val capacity: Int? = null,
+  @Json(name = "building")
+  val building: String? = null,
+  @Json(name = "floor_name")
+  val floorName: String? = null,
+  @Json(name = "floor_section")
+  val floorSection: String? = null,
+  @Json(name = "floor_number")
+  val floorNumber: Int? = null,
 )